extract component in their own module
