+++ gb-bus/src/address_bus.rs
@@ -1,3 +1,9 @@
+pub mod error;
+pub mod iter;
+
+pub use error::Error;
+use iter::Iter;
+
 use crate::FileOperation;
 use crate::{Address, Area};
 
@@ -182,44 +188,4 @@ mod test_address_bus {
         assert_eq!(addr_bus.read(0xff80), Ok(0x37));
         assert_eq!(addr_bus.read(0xffff), Ok(0x38));
     }
-}
-
-pub struct Iter<'a> {
-    current_address: u16,
-    stop: bool,
-    bus: &'a AddressBus,
-}
-
-impl<'a> Iter<'a> {
-    fn new(bus: &'a AddressBus) -> Self {
-        Self {
-            current_address: 0,
-            stop: false,
-            bus,
-        }
-    }
-}
-
-impl<'a> Iterator for Iter<'a> {
-    type Item = u8;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.stop {
-            let bit = self.bus.read(self.current_address).ok();
-            match self.current_address {
-                0xfea0..=0xfeff => self.current_address = 0xff00,
-                0xffff => self.stop = true,
-                _ => self.current_address += 1,
-            }
-            bit
-        } else {
-            None
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub enum Error {
-    BusError(Address),
-    SegmentationFault(Address),
 }
